backend/python/crawling/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
jinja2==3.1.2
bs4==0.0.1
fake_useragent==1.2.1
pandas==2.1.0
tqdm==4.66.1
httpx==0.24.1
motor==3.3.1
python_dotenv==1.0.0
pymongo==4.5.0
apscheduler==3.10.4
